Possible Return?
It's possible.. Nothing here yet though as I'm trying things out and testing the waters, it's been a very long time since I last messed with javascript so I have to catch up on things, and figure out all the changes to adventure.land before I even commit anything worthwhile here.

The master branch is going to be frozen in time and won't really be touched other than to fix the ajax links due to rawgit shutting down, anything new will be commited here for the time (once I actually start committing stuff that is) being until I figure out a different name for a branch and then switch over to that branch and make it the default branch.